+++ CHANGELOG.rst
@@ -12,7 +12,8 @@ Changed
   - Change the format of ``dict`` representation of enumerators to
     have more explicit ordering of members
   - Improvements to the documentation
-  - Use Python enums to enumerate the possible primary types in EnumECG
+  - Use Python enums to enumerate the possible primary types and documentation
+    styles in the EnumECG library
 
 Version 0.3
 -----------